requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0+cpu
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.0.0
scipy>=1.7.0
plotly>=5.14.0
mysql-connector-python>=8.0.0
fastapi>=0.115.0
uvicorn>=0.30.0,<0.34.0
python-multipart>=0.0.20
